Add feedupdater command
